+++ lib/ui/pages/libras/widgets/libras_custom_search_bar.dart
